refactor: consolidate webview helpers
