--- src/common/modules/vscodeApi.js
@@ -1,2 +1,0 @@
-// Initialize VS Code API
-export const vscode = acquireVsCodeApi();
