+++ src/historyView/index.html
@@ -19,8 +19,7 @@
       {
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
-          "@common/vscodeApi.js": "${vscodeApiUri}",
-          "@common/messageRouter.js": "${messageRouterUri}"
+          "@common/webviewContext.js": "${webviewContextUri}"
         }
       }
     </script>
