+++ src/progressView/index.html
@@ -23,12 +23,11 @@
           "./modules/domHandlers.js": "${domHandlersUri}",
           "./modules/constants.js": "${constantsUri}",
           "./modules/logFormatters.js": "${logFormattersUri}",
-          "@common/vscodeApi.js": "${vscodeApiUri}",
+          "@common/webviewContext.js": "${webviewContextUri}",
           "@common/webviewState.js": "${webviewStateUri}",
           "@common/templateUtils.js": "${templateUtilsUri}",
           "@common/domUtils.js": "${domUtilsUri}",
-          "@common/stringUtils.js": "${stringUtilsUri}",
-          "@common/messageRouter.js": "${messageRouterUri}"
+          "@common/stringUtils.js": "${stringUtilsUri}"
         }
       }
     </script>
