+++ src/progressView/modules/domHandlers.js
@@ -1,4 +1,4 @@
-import { vscode } from '@common/vscodeApi.js';
+import { vscode } from '@common/webviewContext.js';
 import {
   getCurrentStream,
   setStreamStatus,
