+++ src/webview/WebviewContentProvider.ts
@@ -45,8 +45,7 @@ export class WebviewContentProvider {
       const templateUtilsUri = getCommonUri('modules/templateUtils.js');
       const domUtilsUri = getCommonUri('modules/domUtils.js');
       const stringUtilsUri = getCommonUri('modules/stringUtils.js');
-      const messageRouterUri = getCommonUri('modules/messageRouter.js');
-      const vscodeApiUri = getCommonUri('modules/vscodeApi.js');
+      const webviewContextUri = getCommonUri('modules/webviewContext.js');
 
       const codiconUri = getNodeModulesUri('@vscode/codicons/dist/codicon.css');
       const codiconsFontUri = getNodeModulesUri(
@@ -78,8 +77,7 @@ export class WebviewContentProvider {
         fileHandlersUri,
         uiHandlersUri,
         templateUtilsUri,
-        messageRouterUri,
-        vscodeApiUri,
+        webviewContextUri,
         codiconUri,
         codiconsFontUri,
       });
