+++ src/webview/index.html
@@ -17,12 +17,11 @@
           "./modules/fileHandlers.js": "${fileHandlersUri}",
           "./modules/uiHandlers.js": "${uiHandlersUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
-          "@common/vscodeApi.js": "${vscodeApiUri}",
+          "@common/webviewContext.js": "${webviewContextUri}",
           "@common/webviewState.js": "${webviewStateUri}",
           "@common/templateUtils.js": "${templateUtilsUri}",
           "@common/domUtils.js": "${domUtilsUri}",
           "@common/stringUtils.js": "${stringUtilsUri}",
-          "@common/messageRouter.js": "${messageRouterUri}",
           "sortablejs": "https://cdn.skypack.dev/sortablejs"
         }
       }
